tlscontext: CC compliant context validation
In order to be Common Criteria compliant, we have to validate each
certificate in the trust chain. As a consequence, we can not ignore
missing CRLs and mismatching purpose errors, however, we still want to
present sensible error messages.

A better (and from the upstream's point of view, more
backwards-compatible) solution would have been to create new
configuration options for handling SSL errors, but that would have
required more code change in syslog-ng than we were keen on doing.